Fixed bug #943.  I was forcing the time to an int.  Now it is being forced to a float.
git-svn-id: https://svn.hugllc.com/Timeclock/trunk@1345 b157844d-9f31-0410-aa6b-d326cbe640ab

+++ ComTimeclock/site/models/timeclock.php
@@ -594,10 +594,10 @@ class TimeclockModelTimeclock extends JModel
         $date = JRequest::getVar('date', '', '', 'string');
         $user =& JFactory::getUser();        
         if (empty($date)) return false;
-        
+
         $ret = true;
         foreach ($timesheet as $data) {
-            $data["hours"] = (int) $data["hours"];
+            $data["hours"] = (float) $data["hours"];
             // If there are no hours don't create a record.
             // If there is already a record allow 0 hours.
             if (empty($data["hours"]) && empty($data["id"])) continue;
